Let manager veto in-session upgrades
A veto will keep the glue scope with all current requires alive. The
result is that the environment kind of becomes a Frankenstein monster at
this point.
The manager will not start any new downloads when this happens and kill
all resumable downloads. When the non-resumable downloads finish, the
manager will "reboot".