Change "cmd + i" shortcut to Get Info
"cmd + i" is generally the standard for opening an information pane, at least
on Mac OS. I think that it is beneficial that such a commonly used command to
be more intuitive than "alt + i". Because of this change, Invert Selection was
changed to "cmd + shift + I" which I personally think is fair.